+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.1.1
+
+- Add an experimental flag to `NgParser` (`toolFriendlyAstOrigin`) which wraps
+  de-sugared AST origins in another synthetic AST that represents the
+  intermediate value (i.e. `BananaAst` or `StarAst`)
+
 ## 0.1.0
 
 - Initial commit